Move types to a separate file
