+++ packages/config/src/layer2s/common/stages/stage.ts
@@ -1,41 +1,10 @@
-export type StageBlueprint = Record<
-  string,
-  {
-    name: string
-    items: Record<
-      string,
-      {
-        positive: string
-        negative: string
-      }
-    >
-  }
->
-
-export type ChecklistTemplate<T extends StageBlueprint> = {
-  [K in keyof T]: {
-    [L in keyof T[K]['items']]: boolean | null | [boolean, string]
-  }
-}
-
-export interface StageSummary {
-  stage: string
-  requirements: {
-    satisfied: boolean
-    description: string
-  }[]
-}
-
-interface MissingStageRequirements {
-  nextStage: string
-  requirements: string[]
-}
-
-export interface StageConfig {
-  stage: string | undefined
-  missing?: MissingStageRequirements
-  summary: StageSummary[]
-}
+import {
+  ChecklistTemplate,
+  MissingStageRequirements,
+  StageBlueprint,
+  StageConfig,
+  StageSummary,
+} from './types'
 
 export function createGetStage<T extends StageBlueprint>(
   blueprint: T,
